test: Refactored unittests accordingly
